implement GGXEVMCurrencyAdapter for currency
